smashgg: Use participant tag rather than player tag
This is mainly relevant when loading past tournaments; we want to load
the name they used to enter the tournament, rather than their current
name.

The current prefix is still used as a fallback; we may want to revisit
this in the future. Added some comments explainin why the code works
this way.