Upload: Disallow negative timestamp offsets in group vods
Previous behavior would result in 23:mm:ss timestamps, but an actual
negative timestamp wouldn't be useful either. Set timestamps before the
start of a group will now be output as 00:00:00.

Also use the set start time directly instead of rounding to nearest
keyframe. Not sure why I did that in the first place.